Extend the non-blocking interface to support CUDA streams
RequestType is now a template parameter, which is int for MPI
backend, and cudaStream_t for NCCL and MPI-CUDA. Wait and Test are
also changed to template functions accordingly. NCCL and MPI-CUDA now
work with the non-blocking API as well, although I have not
evaluated its overlapping effectiveness.

For NCCL and MPI-CUDA, they use an internal default stream when
blocking API is used. The user code can pass a stream as a
request parameter, which can be subsequently used with Wait and Test.

One caveat is that MPI-CUDA does not work with
test_multi_nballreduces as it internally uses a temporary
workspace. If multiple calls are made, they will conflict when using
the workspace buffer. This can be avoided by managing different
workspace buffers for different streams, but its memory
consumption might become a problem. As I am not sure how
important to support multiple asynchronous allreduces, I'll leave it as
it is.