Fix start order issue in host-transfer allreduce.
CUDA can schedule streams and callbacks in an arbitrary order, which leads to allreduces starting at different times, which leads to deadlock. Current fix is to run all non-blocking host-transfer allreduces on a single stream. This hurts performance some and leads to a potential issue with blocking host-transfer allreduces leading to deadlock.