Reuse CUDA memories in benchmark_allreduce
Spectrum MPI on Sierra seems to have a large initialization overhead for
a GPU buffer, which significantly impacted the benchmark performance of
the MPI-CUDA ring allreduce. Since GPU memories should be reused in
common cases without frequent malloc/free, it should be reasonable to
assume same memory chunks are repeated used in applications.