Add segmented allreduce benchmark
This adds a benchmark for segmented allreduces. Among its command line
arguments, it takes a file to specify the sizes; this is to ease its
use in benchmarking many configurations with a single run, since
jsrun startup time becomes a significant overhead otherwise.

It can also benchmark segmented allreduces based on whether the
segmenting is done via mod or not. Our current use cases require such
a distribution, but varying this can help identify interference
effects.

This also adjusts the code for printing out benchmark results. When
passing in the communicator size, it now expects the size to be for a
single segment. This simplifies user code.